Fix transaction service HTTP 400 errors by switching DTOs from BigDecimal to Double
- Updated all transaction DTOs (CreateTransactionRequest, DepositRequest, WithdrawRequest, TransferRequest) to use Double instead of BigDecimal for amount fields
- Modified validation annotations from @DecimalMin to @Positive for better compatibility
- Added BigDecimal.valueOf() conversion in service layer to maintain precision for database operations
- Enhanced error handling in TransactionController with BindingResult validation
- Added GlobalExceptionHandler for better validation error reporting
- Updated AccountsServiceClient to use HttpServletRequest instead of raw JWT tokens
- Synchronized transaction methods to prevent race conditions

This resolves the JSON deserialization issues that were causing HTTP 400 errors when the user service attempted to create transactions during demo data generation.

+++ backend/transactions-service/src/main/java/com/banking/transactionsservice/dto/DepositRequest.java
@@ -3,7 +3,6 @@ package com.banking.transactionsservice.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
-import java.math.BigDecimal;
 
 public class DepositRequest {
     
@@ -12,13 +11,13 @@ public class DepositRequest {
     
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be positive")
-    private BigDecimal amount;
+    private Double amount;
     
     private String description;
     
     public DepositRequest() {}
     
-    public DepositRequest(Long accountId, BigDecimal amount, String description) {
+    public DepositRequest(Long accountId, Double amount, String description) {
         this.accountId = accountId;
         this.amount = amount;
         this.description = description;
@@ -32,11 +31,11 @@ public class DepositRequest {
         this.accountId = accountId;
     }
     
-    public BigDecimal getAmount() {
+    public Double getAmount() {
         return amount;
     }
     
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
